move unstable_components to internal
